requirements.txt
astor==0.8.1
astroid==2.15.3
asttokens==2.1.0
backcall==0.2.0
bcrypt==3.2.0
binarytree==6.5.1
blinker==1.4
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
cmake==3.24.1.1
colorama==0.4.5
contourpy==1.0.5
cryptography==3.4.8
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.6
distlib==0.3.5
distro==1.7.0
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
face-recognition-models==0.3.0
fasteners==0.17.3
filelock==3.7.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-MySQL==1.5.2
flynt==0.77
fonttools==4.37.4
future==0.18.2
graphviz==0.20.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.16.2
ipython==8.6.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.2
jupyter_client==7.4.4
jupyter_core==4.11.2
keyring==23.8.2
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
libevdev==0.5
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pylint==2.17.2
pymacaroons==0.13.0
pymongo==4.3.3
PyMySQL==1.0.3
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
requests==2.28.2
SecretStorage==3.3.3
setuptools-scm==7.1.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.7
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
